+++ TombLib/TombLib/LevelData/RoomGeometry/Room2DPoint.cs
@@ -1,0 +1,13 @@
+﻿namespace TombLib.LevelData;
+
+public struct Room2DPoint
+{
+	public int X;
+	public int Z;
+
+	public Room2DPoint(int x, int z)
+	{
+		X = x;
+		Z = z;
+	}
+}
