+++ TombLib/TombLib/LevelData/RoomGeometry/WallEnd.cs
@@ -1,0 +1,24 @@
+﻿namespace TombLib.LevelData;
+
+public struct WallEnd
+{
+	/// <summary>
+	/// X coordinate of the wall point.
+	/// </summary>
+	public int X;
+
+	/// <summary>
+	/// Z coordinate of the wall point.
+	/// </summary>
+	public int Z;
+
+	/// <summary>
+	/// Minimum Y coordinate (height) of the wall point.
+	/// </summary>
+	public int MinY;
+
+	/// <summary>
+	/// Maximum Y coordinate (height) of the wall point.
+	/// </summary>
+	public int MaxY;
+}
