+++ TombLib/TombLib/LevelData/RoomGeometry/WallSplit.cs
@@ -1,0 +1,13 @@
+﻿namespace TombLib.LevelData;
+
+public struct WallSplit
+{
+	public int StartY;
+	public int EndY;
+
+	public WallSplit(int startY, int endY)
+	{
+		StartY = startY;
+		EndY = endY;
+	}
+}
